requirements.txt
Flask=3.0.3
Flask-JWT-Extended==4.7.1
Flask-WTF==1.2.2
Flask-SQLAlchemy ==3.1.1
argon2-cffi==25.1.0
